# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Open-source drug information gateway. Provides NDC lookup, therapeutic class search, drug interactions, RxNorm fuzzy search, and structured product label data — all through a clean REST API with Redis caching, per-key rate limiting, and circuit breaker protection.\n\n## Getting Started\n1. **Create an API key:** `POST /admin/keys` with `Authorization: Bearer {ADMIN_SECRET}` and body `{\"app_name\": \"my-app\", \"origins\": [\"*\"], \"rate_limit\": 1000}`\n2. **Search for a drug:** `GET /v1/drugs/autocomplete?q=lipit` with header `X-API-Key: {your_key}` — returns matching drug names\n3. **Look up details:** `GET /v1/drugs/class?name=atorvastatin` — returns therapeutic classes and brand names\n4. **Get interactions:** `GET /v1/drugs/info?name=warfarin` — returns FDA label interaction warnings\n5. **Check multi-drug interactions:** `POST /v1/drugs/interactions` with `{\"drugs\": [{\"name\": \"warfarin\"}, {\"name\": \"aspirin\"}]}`",
        "title": "drug-gate API",
        "contact": {
            "name": "drug-gate on GitHub",
            "url": "https://github.com/finish06/drug-gate"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/finish06/drug-gate/blob/main/LICENSE"
        },
        "version": "0.9.0"
    },
    "basePath": "/",
    "paths": {
        "/admin/cache": {
            "delete": {
                "security": [
                    {
                        "AdminAuth": []
                    }
                ],
                "description": "Deletes all cache keys matching cache:*, or a subset matching cache:{prefix}* if the prefix query parameter is provided. Uses SCAN to iterate keys safely without blocking Redis. Does not affect API key or rate limit data stored under other key namespaces.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Clear Redis cache",
                "parameters": [
                    {
                        "type": "string",
                        "example": "rxnorm",
                        "description": "Key prefix filter (e.g. rxnorm, drugnames)",
                        "name": "prefix",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.cacheClearResult"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid admin bearer token",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Redis unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/drugs/autocomplete": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns drug names matching the given prefix, optimized for typeahead search UIs. Uses an in-memory prefix index for sub-millisecond response times. The prefix must be at least 2 characters. Results include both generic and brand names with their type labels.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drugs"
                ],
                "summary": "Drug name autocomplete",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ator",
                        "description": "Prefix to match (min 2 chars)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "Max results (default: 10, max: 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugNameEntry"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing or too-short q parameter",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/class": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Looks up a drug by generic or brand name and returns its pharmacological classes (EPC, MoA, PE, CS), deduplicated brand names, and generic name. Tries generic name first and falls back to brand name search. Use this endpoint to discover what class a drug belongs to.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drugs"
                ],
                "summary": "Look up drug class by name",
                "parameters": [
                    {
                        "type": "string",
                        "example": "simvastatin",
                        "description": "Drug name (generic or brand)",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugClassResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or empty name parameter",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Drug not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/drugs/ndc/{ndc}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts a product NDC (dash-separated) and returns drug name, generic name, and therapeutic classes from the FDA NDC Directory. Supports 5-4, 4-4, and 5-3 formats with automatic fallback to zero-padded 5-4 form. Use this endpoint when you have an NDC and need to identify the drug.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drugs"
                ],
                "summary": "Look up drug by NDC",
                "parameters": [
                    {
                        "type": "string",
                        "example": "00069-3150",
                        "description": "Product NDC with dash",
                        "name": "ndc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid NDC format",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No drug found for this NDC",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_finish06_drug-gate_internal_model.DrugClass": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "HMG-CoA Reductase Inhibitor"
                },
                "type": {
                    "type": "string",
                    "example": "EPC"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.DrugClassResponse": {
            "type": "object",
            "properties": {
                "brand_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Zocor"
                    ]
                },
                "classes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.DrugClass"
                    }
                },
                "generic_name": {
                    "type": "string",
                    "example": "SIMVASTATIN"
                },
                "query_name": {
                    "type": "string",
                    "example": "simvastatin"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.DrugDetailResponse": {
            "type": "object",
            "properties": {
                "classes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "HMG-CoA Reductase Inhibitor"
                    ]
                },
                "generic_name": {
                    "type": "string",
                    "example": "ATORVASTATIN CALCIUM"
                },
                "name": {
                    "type": "string",
                    "example": "Lipitor"
                },
                "ndc": {
                    "type": "string",
                    "example": "00069-3150"
                }
            }
        },
...
        "github_com_finish06_drug-gate_internal_model.DrugNameEntry": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Lipitor"
                },
                "type": {
                    "type": "string",
                    "example": "brand"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "message": {
                    "type": "string",
                    "example": "No drug found for NDC 99999-9999"
                }
            }
        },
...
        "internal_handler.cacheClearResult": {
            "type": "object",
            "properties": {
                "keys_deleted": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token set via `ADMIN_SECRET` environment variable. Format: `Bearer {secret}`. Required for all `/admin/*` endpoints.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Publishable API key for frontend applications. Create one via `POST /admin/keys`. Include in all `/v1/*` requests.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    },
...
}